requirements.txt
Django==3.2.3
django-rest-framework==0.1.0